requirements.txt
Flask==1.1.2
google-cloud-bigquery==2.10.0
google-cloud-storage==1.36.1
